36a21724-67fc-471e-a3e5-dbc58111d898

+++ src/main.rs
@@ -5,10 +5,10 @@ use git_history_by_example::custom_add;
 #[command(author, version, about, long_about = None)]
 struct Args {
     /// First number to add
-    first: i64,
+    first: f64,
 
     /// Second number to add
-    second: i64,
+    second: f64,
 }
 
 fn main() {
